+++ packages/tsargp/test/styles/styles.messages.spec.ts
@@ -1,0 +1,97 @@
+import { afterEach, describe, expect, it } from 'bun:test';
+import {
+  ErrorItem,
+  AnsiString,
+  AnsiMessage,
+  WarnMessage,
+  ErrorMessage,
+  TextMessage,
+} from '../../src/library';
+
+process.env['FORCE_WIDTH'] = '0'; // omit styles
+
+describe('AnsiMessage', () => {
+  afterEach(() => {
+    ['NO_COLOR', 'FORCE_COLOR'].forEach((key) => delete process.env[key]);
+    process.env['FORCE_WIDTH'] = '0'; // omit styles
+  });
+
+  it('wrap the message while respecting the environment configuration', () => {
+    const str = new AnsiString().split('type script');
+    const msg = new AnsiMessage(str);
+    expect(msg.wrap(0)).toEqual('type script');
+    expect(msg.wrap(11)).toEqual('type script');
+    process.env['NO_COLOR'] = '1';
+    expect(msg.wrap(0)).toEqual('type script');
+    expect(msg.wrap(11)).toEqual('type script');
+    process.env['FORCE_COLOR'] = '1';
+    expect(msg.wrap(0)).toEqual('type script');
+    expect(msg.wrap(11)).toEqual('type script');
+    process.env['FORCE_WIDTH'] = '10';
+    expect(msg.message).toEqual('type\nscript');
+  });
+
+  it('produce a string message', () => {
+    const str = new AnsiString().split('type script');
+    const msg = new AnsiMessage(str);
+    expect(msg.message).toEqual('type script');
+  });
+});
+
+describe('WarnMessage', () => {
+  afterEach(() => {
+    process.env['FORCE_WIDTH'] = '0'; // omit styles
+  });
+
+  it('wrap the message while respecting the environment configuration', () => {
+    const str = new AnsiString().split('type script');
+    const msg = new WarnMessage(str);
+    process.env['FORCE_WIDTH'] = '10';
+    expect(msg.message).toEqual('type\nscript');
+  });
+
+  it('produce a string message', () => {
+    const str = new AnsiString().split('type script');
+    const msg = new WarnMessage(str);
+    expect(msg.message).toEqual('type script');
+  });
+
+  describe('add', () => {
+    it('format a string message with an error phrase', () => {
+      const msg = new WarnMessage();
+      msg.add(ErrorItem.missingRequiredOption, {}, 'abc');
+      expect(msg.message).toEqual(`Option 'abc' is required.\n`);
+    });
+  });
+
+  describe('addCustom', () => {
+    it('format a string message with a custom phrase', () => {
+      const msg = new WarnMessage();
+      msg.addCustom('#0 #1 #2', {}, 0, 'abc', false);
+      expect(msg.message).toEqual(`0 'abc' false\n`);
+    });
+  });
+});
+
+describe('ErrorMessage', () => {
+  describe('create', () => {
+    it('format a string message with an error phrase', () => {
+      const msg = ErrorMessage.create(ErrorItem.missingRequiredOption, {}, 'abc');
+      expect(msg.message).toEqual(`Option 'abc' is required.\n`);
+    });
+  });
+
+  describe('createCustom', () => {
+    it('format a string message with a custom phrase', () => {
+      const msg = ErrorMessage.createCustom('#0 #1 #2', {}, 0, 'abc', false);
+      expect(msg.message).toEqual(`0 'abc' false\n`);
+    });
+  });
+});
+
+describe('TextMessage', () => {
+  it('produce a string message with line feeds', () => {
+    const msg = new TextMessage('type', 'script');
+    expect(msg.message).toEqual('type\nscript');
+  });
+});
